\name{import_gentry}
\alias{import_gentry}

\title{Smart Import of Gentry's Transect Data.}
\description{
Import and pre-format of data obtained from the \bold{Alwyn H. Gentry Forest
Transect Data Set} (\url{http://www.mobot.org/mobot/research/gentry/welcome.shtml}).
}
\usage{
import_gentry(path, ...)
}
\arguments{
\item{path}{Folder containing the down loaded tables.}
\item{...}{Further arguments passed to \code{\link{read_xls}}.}
}
\details{
Documentation under construction
}
\value{
Documentation under construction
}
\author{
Miguel Alvarez, \email{kamapu78@gmail.com}.
}
\examples{
library(taxsearch)
## Documentation under construction
}
