\name{tropicos_specimens}
\alias{tropicos_specimens}

\title{Query for Specimens in \dQuote{Tropicos}.}
\description{
Documentation under construction
}
\usage{
tropicos_specimens(key, collector, number, specimen, mode="collector",
                 encoding="UTF-8")
}
\arguments{
\item{key}{API key.}
\item{collector}{Name of senior collector.}
\item{number}{Collection number by respective senior collector.}
\item{specimen}{Specimen ID.}
\item{mode}{Entry used to retrieve specimen.}
\item{encoding}{Character encoding.}
}
\details{
Documentation under construction
}
\value{
Documentation under construction
}
\author{
Miguel Alvarez, \email{kamapu78@gmail.com}.
}
\examples{
library(taxsearch)
## Documentation under construction
}
